+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased] - 2025-09-16
 ### Added
 - Audio annotations if audio and video synchronized
+- Process sound for waveforms and spectrogram
 
 ## [2.3.0] - 2025-09-13
 ### Added